+++ app/src/androidTest/java/nextstep/signup/SignupTextFieldValidationTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import androidx.compose.ui.test.performTextInput
-import nextstep.signup.ui.SignupScreen
+import nextstep.signup.ui.signin.SignupScreen
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
