+++ app/src/main/java/nextstep/signup/MainActivity.kt
@@ -3,7 +3,7 @@ package nextstep.signup
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import nextstep.signup.ui.SignupScreen
+import nextstep.signup.ui.signin.SignupScreen
 import nextstep.signup.ui.theme.SignupTheme
 
 class MainActivity : ComponentActivity() {
